requirements.txt
mistralai==1.2.2
python-dotenv==1.0.1
requests==2.32.3
